refactor: Extract Preview Block from SelectionLevelInfo into separate class
